+++ src/main/java/com/pikapedia/account/MyRewardC.java
@@ -1,4 +1,4 @@
-package com.pikapedia.search;
+package com.pikapedia.account;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,8 +6,6 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import com.pikapedia.account.AccountDAO;
 
 
 @WebServlet("/MyRewardC")
