+++ src/main/java/com/pikapedia/main/HC.java
@@ -8,13 +8,13 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.pikapedia.account.AccountDAO;
-import com.pikapedia.search.SearchDAO;
+import com.pikapedia.db.DBDAO;
 
 @WebServlet("/HC")
 public class HC extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		AccountDAO.checkLogin(request);
-		SearchDAO.getAllPokemon(request);
+		DBDAO.getAllPokemon(request);
 		request.setAttribute("contentPage", "jsp/pokemonMain.jsp");
 		request.getRequestDispatcher("index.jsp").forward(request, response);
 	}
