+++ src/main/webapp/css/myRewards.css
@@ -147,35 +147,3 @@
 	color: #000000;
 }
 
-/* slider */
-.awards .swiper-prev,
-.awards .swiper-next {
-  width: 42px;
-  height: 42px;
-  border: 2px solid #fff;
-  border-radius: 50%;
-  outline: none;
-  color: #fff;
-  opacity: 0.3;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  color: #000000;
-}
-
-.awards .swiper-prev {
-  left: -100px;
-}
-
-.awards .swiper-next {
-  right: -100px;
-}
-.awards .swiper-prev:hover,
-.awards .swiper-next:hover {
-  background-color: #fff;
-  color: #333;
-}
