+++ src/main/webapp/css/pokedexFrameAccount.css
@@ -42,7 +42,7 @@
 }
 
 #pokedex-blue-frame {
-/*   box-sizing: border-box;
+/*box-sizing: border-box;
   position: absolute;
   top: 0;
   right: 0;
