--- src/main/webapp/jsp/pokemonDetail.jsp
@@ -1,14 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>pokemonDetail</title>
-
-</head>
-<body>
-	<h1>Pokemon Detail Page</h1>
-    
-</body>
-</html>
