+++ src/main/webapp/jsp/quizOutput.jsp
@@ -1,4 +1,4 @@
-<%@page import="com.pikapedia.search.Pokemon"%>
+<%@page import="com.pikapedia.db.Pokemon"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <!DOCTYPE html>
 <html>
